Create README.md
I am a research scholar in English literature. My PhD topic is about using film adaptations of literary texts as a pedagogical aid to teach English in the English literature M.A. classroom. I am also thinking about how to integrate film as a pedagogical aid in the classroom, as the curriculums are jam packed with many other activities , and limited number of classes allocated to teachers.